feat: update step2
